Backend/requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
google-generativeai==0.8.5
fastapi
uvicorn